requirements.txt
Django
django-rest-framework
psycopg2-binary
python-decouple
envparse
